Extract timedDelay out to internal function
